+++ src/main/java/com/alevnarcin/librarymanagementsystem/service/BorrowedService.java
@@ -6,11 +6,13 @@ import com.alevnarcin.librarymanagementsystem.dto.BorrowedDto;
 import com.alevnarcin.librarymanagementsystem.entity.BookEntity;
 import com.alevnarcin.librarymanagementsystem.entity.BorrowedEntity;
 import com.alevnarcin.librarymanagementsystem.entity.PersonEntity;
+import com.alevnarcin.librarymanagementsystem.exception.response.ExceptionResponse;
 import com.alevnarcin.librarymanagementsystem.repository.BookRepository;
 import com.alevnarcin.librarymanagementsystem.repository.BorrowedRepository;
 import com.alevnarcin.librarymanagementsystem.repository.PersonRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.xml.crypto.Data;
 import java.awt.print.Book;
@@ -26,35 +28,50 @@ public class BorrowedService {
     private final BookRepository bookRepository;
     private final PersonRepository personRepository;
 
+    @Transactional
     public BorrowedDto find(Integer id) {
-        BorrowedEntity borrowedEntity = borrowedRepository.findById(id).orElseThrow(NoSuchElementException::new);
+        BorrowedEntity borrowedEntity = borrowedRepository.findById(id).orElse(null);
+        if (borrowedEntity == null) {
+            throw new ExceptionResponse("Oopps cannot find borrowed... ");
+        }
         return borrowedConverter.borrowedEntityToBorrowedDto(borrowedEntity);
     }
 
+    @Transactional
     public BorrowedDto create(BorrowedDto borrowedDto) {
         BorrowedEntity borrowedEntity = borrowedConverter.borrowedDtoToBorrowedEntity(borrowedDto);
         BorrowedEntity savedBorrowedEntity = borrowedRepository.save(borrowedEntity);
+
         return borrowedConverter.borrowedEntityToBorrowedDto(savedBorrowedEntity);
     }
 
+    @Transactional
     public BorrowedDto update(BorrowedDto borrowedDto, Integer borrowedId) {
-        BorrowedEntity borrowedEntity = borrowedRepository.findById(borrowedId).orElseThrow(NoSuchElementException::new);
+        BorrowedEntity borrowedEntity = borrowedRepository.findById(borrowedId).orElse(null);
+        if (borrowedEntity == null) {
+            throw new ExceptionResponse("Oopps cannot find borrowed... ");
+        }
 
-        borrowedEntity.setId(borrowedDto.getId());
-        borrowedEntity.setBorrowStartDate(borrowedDto.getBorrowStartDate());
-        //borrowedEntity.setBorrowEndDate(borrowedDto.getBorrowEndDate());
-        //borrowedEntity.setBorrowReturnDate(borrowedDto.getBorrowReturnDate());
+        borrowedEntity.setBorrowReturnDate(borrowedDto.getBorrowReturnDate());
 
         return borrowedConverter.borrowedEntityToBorrowedDto(borrowedRepository.save(borrowedEntity));
     }
 
+    @Transactional
     public void delete(Integer borrowedId) {
-        BorrowedEntity borrowedEntity = borrowedRepository.findById(borrowedId).orElseThrow(NoSuchElementException::new);
+        BorrowedEntity borrowedEntity = borrowedRepository.findById(borrowedId).orElse(null);
+        if (borrowedEntity == null) {
+            throw new ExceptionResponse("Oopps cannot find borrowed... ");
+        }
         borrowedRepository.delete(borrowedEntity);
     }
 
-    public BookEntity getBook(Integer borrowedId, Integer bookId) {
-        BookEntity bookEntity = bookRepository.findById(bookId).orElseThrow(NoSuchElementException::new);
+    @Transactional
+    public BookEntity getBook(Integer bookId) {
+        BookEntity bookEntity = bookRepository.findById(bookId).orElse(null);
+        if (bookEntity == null) {
+            throw new ExceptionResponse("Oopps cannot find book... ");
+        }
         BorrowedEntity borrowedEntity = new BorrowedEntity();
         borrowedEntity.setBookEntity(bookEntity);
         borrowedEntity.setBorrowStartDate(LocalDate.now());
@@ -63,11 +80,17 @@ public class BorrowedService {
         return bookRepository.save(bookEntity);
     }
 
+    @Transactional
     public PersonEntity getBorrow(Integer bookId, Integer personId) {
-        PersonEntity personEntity = personRepository.findById(personId).orElseThrow(NoSuchElementException::new);
-        BookEntity bookEntity = bookRepository.findById(bookId).orElseThrow(NoSuchElementException::new);
+        PersonEntity personEntity = personRepository.findById(personId).orElse(null);
+        if (personEntity == null) {
+            throw new ExceptionResponse("Oopps cannot find person... ");
+        }
+        BookEntity bookEntity = bookRepository.findById(bookId).orElse(null);
+        if (bookEntity == null) {
+            throw new ExceptionResponse("Oopps cannot find book... ");
+        }
         BorrowedEntity borrowedEntity = new BorrowedEntity();
-
         borrowedEntity.setPersonEntity(personEntity);
         borrowedEntity.setBookEntity(bookEntity);
         borrowedEntity.setBorrowStartDate(LocalDate.now());
